Moved ConnectFragment to ui folder for consistency
